Add base version to play-ws-common javaagent module
- instrumentation/play/play-ws/play-ws-common -> play-ws-common-1.0
  (package io.opentelemetry.javaagent.instrumentation.playws
   -> io.opentelemetry.javaagent.instrumentation.playws.common.v1_0,
    in both javaagent and testing source sets)
