+++ instrumentation/play/play-ws/play-ws-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/playws/v1_0/PlayWs10Singletons.java
@@ -6,7 +6,7 @@
 package io.opentelemetry.javaagent.instrumentation.playws.v1_0;
 
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.javaagent.instrumentation.playws.PlayWsClientInstrumenterFactory;
+import io.opentelemetry.javaagent.instrumentation.playws.common.v1_0.PlayWsClientInstrumenterFactory;
 import play.shaded.ahc.org.asynchttpclient.Request;
 import play.shaded.ahc.org.asynchttpclient.Response;
