+++ instrumentation/play/play-ws/play-ws-1.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/playws/v1_0/PlayScalaWsClientTest.java
@@ -5,6 +5,6 @@
 
 package io.opentelemetry.javaagent.instrumentation.playws.v1_0;
 
-import io.opentelemetry.javaagent.instrumentation.playws.PlayScalaWsClientBaseTest;
+import io.opentelemetry.javaagent.instrumentation.playws.common.v1_0.PlayScalaWsClientBaseTest;
 
 class PlayScalaWsClientTest extends PlayScalaWsClientBaseTest {}
