+++ instrumentation/play/play-ws/play-ws-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/playws/v2_0/PlayWsInstrumentationModule.java
@@ -13,9 +13,9 @@ import com.google.auto.service.AutoService;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.instrumentation.playws.AbstractBootstrapInstrumentation;
-import io.opentelemetry.javaagent.instrumentation.playws.AsyncHttpClientInstrumentation;
-import io.opentelemetry.javaagent.instrumentation.playws.HandlerPublisherInstrumentation;
+import io.opentelemetry.javaagent.instrumentation.playws.common.v1_0.AbstractBootstrapInstrumentation;
+import io.opentelemetry.javaagent.instrumentation.playws.common.v1_0.AsyncHttpClientInstrumentation;
+import io.opentelemetry.javaagent.instrumentation.playws.common.v1_0.HandlerPublisherInstrumentation;
 import java.util.List;
 import javax.annotation.Nullable;
 import net.bytebuddy.asm.Advice;
