+++ instrumentation/play/play-ws/play-ws-2.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/playws/v2_1/PlayJavaWsClientTest.java
@@ -5,6 +5,6 @@
 
 package io.opentelemetry.javaagent.instrumentation.playws.v2_1;
 
-import io.opentelemetry.javaagent.instrumentation.playws.PlayJavaWsClientBaseTest;
+import io.opentelemetry.javaagent.instrumentation.playws.common.v1_0.PlayJavaWsClientBaseTest;
 
 class PlayJavaWsClientTest extends PlayJavaWsClientBaseTest {}
