+++ instrumentation/play/play-ws/play-ws-common-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/playws/common/v1_0/PlayWsClientHttpAttributesGetter.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.playws.common.v1_0;
+
+import io.opentelemetry.instrumentation.api.internal.HttpConstants;
+import io.opentelemetry.instrumentation.api.semconv.http.HttpClientAttributesGetter;
+import java.net.InetSocketAddress;
+import java.util.List;
+import javax.annotation.Nullable;
+import play.shaded.ahc.org.asynchttpclient.Request;
+import play.shaded.ahc.org.asynchttpclient.Response;
+import play.shaded.ahc.org.asynchttpclient.uri.Uri;
+
+final class PlayWsClientHttpAttributesGetter
+    implements HttpClientAttributesGetter<Request, Response> {
+
+  @Override
+  public String getHttpRequestMethod(Request request) {
+    return request.getMethod();
+  }
+
+  @Override
+  public String getUrlFull(Request request) {
+    return request.getUri().toUrl();
+  }
+
+  @Override
+  public List<String> getHttpRequestHeader(Request request, String name) {
+    return request.getHeaders().getAll(name);
+  }
+
+  @Override
+  public Integer getHttpResponseStatusCode(
+      Request request, Response response, @Nullable Throwable error) {
+    return response.getStatusCode();
+  }
+
+  @Override
+  public List<String> getHttpResponseHeader(Request request, Response response, String name) {
+    return response.getHeaders().getAll(name);
+  }
+
+  @Nullable
+  @Override
+  public String getServerAddress(Request request) {
+    return request.getUri().getHost();
+  }
+
+  @Nullable
+  @Override
+  public Integer getServerPort(Request request) {
+    Uri uri = request.getUri();
+    return HttpConstants.portOrDefaultFromScheme(uri.getPort(), uri.getScheme());
+  }
+
+  @Override
+  @Nullable
+  public InetSocketAddress getNetworkPeerInetSocketAddress(
+      Request request, @Nullable Response response) {
+    if (response != null && response.getRemoteAddress() instanceof InetSocketAddress) {
+      return (InetSocketAddress) response.getRemoteAddress();
+    }
+    return null;
+  }
+}
