--- instrumentation/play/play-ws/play-ws-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/playws/AbstractBootstrapInstrumentation.java
@@ -1,52 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.playws;
-
-import static net.bytebuddy.matcher.ElementMatchers.named;
-
-import io.opentelemetry.context.Scope;
-import io.opentelemetry.javaagent.bootstrap.Java8BytecodeBridge;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import javax.annotation.Nullable;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-
-public class AbstractBootstrapInstrumentation implements TypeInstrumentation {
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("play.shaded.ahc.io.netty.bootstrap.AbstractBootstrap");
-  }
-
-  @Override
-  public void transform(TypeTransformer transformer) {
-    transformer.applyAdviceToMethod(
-        named("initAndRegister"), getClass().getName() + "$DisablePropagationAdvice");
-  }
-
-  @SuppressWarnings("unused")
-  public static class DisablePropagationAdvice {
-    @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
-    @Nullable
-    public static Scope onEnter() {
-      // Prevent context from leaking by running this method under root context.
-      // Root context is not propagated by executor instrumentation.
-      if (Java8BytecodeBridge.currentContext() != Java8BytecodeBridge.rootContext()) {
-        return Java8BytecodeBridge.rootContext().makeCurrent();
-      }
-      return null;
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)
-    public static void onExit(@Advice.Enter @Nullable Scope scope) {
-      if (scope != null) {
-        scope.close();
-      }
-    }
-  }
-}
