--- instrumentation/play/play-ws/play-ws-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/playws/HandlerPublisherInstrumentation.java
@@ -1,41 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.playws;
-
-import static net.bytebuddy.matcher.ElementMatchers.named;
-
-import io.opentelemetry.javaagent.bootstrap.Java8BytecodeBridge;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.asm.Advice.AssignReturned;
-import net.bytebuddy.asm.Advice.AssignReturned.ToArguments.ToArgument;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.reactivestreams.Subscriber;
-
-public class HandlerPublisherInstrumentation implements TypeInstrumentation {
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("play.shaded.ahc.com.typesafe.netty.HandlerPublisher");
-  }
-
-  @Override
-  public void transform(TypeTransformer transformer) {
-    transformer.applyAdviceToMethod(
-        named("subscribe"), getClass().getName() + "$WrapSubscriberAdvice");
-  }
-
-  @SuppressWarnings("unused")
-  public static class WrapSubscriberAdvice {
-
-    @AssignReturned.ToArguments(@ToArgument(0))
-    @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
-    public static Subscriber<?> enter(@Advice.Argument(0) Subscriber<?> subscriber) {
-      return new SubscriberWrapper<>(subscriber, Java8BytecodeBridge.currentContext());
-    }
-  }
-}
