--- instrumentation/play/play-ws/play-ws-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/playws/PlayWsClientInstrumenterFactory.java
@@ -1,21 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.playws;
-
-import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.javaagent.bootstrap.internal.JavaagentHttpClientInstrumenters;
-import play.shaded.ahc.org.asynchttpclient.Request;
-import play.shaded.ahc.org.asynchttpclient.Response;
-
-public class PlayWsClientInstrumenterFactory {
-
-  public static Instrumenter<Request, Response> createInstrumenter(String instrumentationName) {
-    return JavaagentHttpClientInstrumenters.create(
-        instrumentationName, new PlayWsClientHttpAttributesGetter(), new HttpHeaderSetter());
-  }
-
-  private PlayWsClientInstrumenterFactory() {}
-}
