+++ settings.gradle.kts
@@ -548,8 +548,8 @@ include(":instrumentation:play:play-mvc:play-mvc-2.6:javaagent")
 include(":instrumentation:play:play-ws:play-ws-1.0:javaagent")
 include(":instrumentation:play:play-ws:play-ws-2.0:javaagent")
 include(":instrumentation:play:play-ws:play-ws-2.1:javaagent")
-include(":instrumentation:play:play-ws:play-ws-common:javaagent")
-include(":instrumentation:play:play-ws:play-ws-common:testing")
+include(":instrumentation:play:play-ws:play-ws-common-1.0:javaagent")
+include(":instrumentation:play:play-ws:play-ws-common-1.0:testing")
 include(":instrumentation:powerjob-4.0:javaagent")
 include(":instrumentation:pulsar:pulsar-2.8:javaagent")
 include(":instrumentation:pulsar:pulsar-2.8:javaagent-unit-tests")
